### 2025-05-29 #8315 @danielskatz
👋 @zhanghe9704 - thanks for your submission. I'm the track editor who will get this started, by finding an editor, after some initial checks.

I have two concerns related to the paper:
1. When you discuss a reference explicitly (by name), you should list it as `@ref` rather than as `[@ref]` which is proper when the reference is discussed implicitly. The first two instances of references in the paper in the background paragraph should be changed. I didn't check this elsewhere.
2. Your paper is about 1750 words long, as JOSS's recommendation is 250-1000 words. So you will need to reduce the length of your paper.  After you make changes, you can use the command `@editorialbot generate pdf` to make a new PDF. And you can check a bunch of things, including the paper's word county, with `@editorialbot check repository`. An editorialbot command needs to be the first entry in a new comment.

Please ping me when you are done with these and ready to proceed.
