### 2025-05-31 0de0ebe
doc: shorten the paper
