+++ paper/paper.md
@@ -21,22 +21,22 @@ bibliography: paper.bib
 
 # Summary
 
-The truncated power series algebra (TPSA), also referred to as differential algebra (DA), is a well-established and widely used method in particle accelerator physics. It is typically used to generate high-order map of a nonlinear dynamical system. The DA calculation is based on the DA vector that can be practically viewed as the Taylor expansion  of a given function at a specific point up to a predetermined order $n$. Besides map generation, the DA/TPSA technique can also be used in symplectic tracking, normal form analysis, verified integration, global optimization, fast multipole method, etc.  This package is the first one to combine the DA/TPSA algorithm with symbolic calculation, and it allows users to carry out symbolic DA-based calculations. The coefficient of any element in a symbolic DA vector is an explicit expression of preselected symbols, in lieu of a number in a numerical DA vector. This capability makes it possible to trace the contribution of initial conditions to the final result in a DA calculation process and improve the efficiency of  repeated DA calculations. It also provides a method to obtain explicit expressions for the higher-order derivatives of an arbitrary function. 
+The truncated power series algebra (TPSA), or differential algebra (DA), is a well-established and widely used method in particle accelerator physics. It is typically used to generate high-order map of a nonlinear dynamical system but is also used in symplectic tracking, normal form analysis, verified integration, global optimization, fast multipole method, etc.  This package is the first  to carry out symbolic DA-based calculations. Symbolic DA makes it possible to trace the contribution of initial conditions to the final result in a DA calculation process and improve the efficiency of  repeated DA calculations, which potentially broadens the usage of DA. 
 
 # Background
 
-In the following, we give a very brief introduction to (symbolic) TPSA/DA from a practical computational perspective. Please refer to [@AIEP108book] and [@chao2002lecture]  for the complete theory with more details.
+We start with a brief introduction to DA from a practical computational perspective. Please refer to @AIEP108book and @chao2002lecture  for the complete theory.
 
 The fundamental concept in DA is the DA vector. To make this concept easier to understand, we can consider a DA vector as the Taylor expansion of a function at a specific point.  
 
 Considering a function $f(\mathbf{x})$ and its Taylor expansion $f_{\mathrm{T}}(\mathbf{x}_0)$  at the point $\mathbf{x}_0$ up to order $n$, we can define  the equivalence relation between the Taylor expansion and the DA vector as follows
 
-$$ [f]_n = f_{\mathrm{T}}(\mathbf{x}_0) = \sum {C_{n_1,n_2, ..., n_v}} \cdot d_1^{n_1} \cdot \dots \cdot d_v^{n_v}, $$ where $\mathbf{x} = (x_1, x_2, \dots, x_v)$, and $n \ge n_1 + n_2 + \dots + n_v$. Here $d_i$ is a special number: it represents a small variance in $x_i$. Generally, one can define a DA vector by directly setting values to respective terms, without defining the function $f$. The addition and multiplication of two DA vectors can be defined straightforwardly. To add two DA vectors, we simply add  the coefficients of the like terms. To multiply two DA vectors, we multiply each term in the first one with all the terms in the second one. We  then combine like terms while ignoring all terms above order $n$. So, given two DA vectors $[a]_n$ and $[b]_n$ and a scalar c, we have the following formulae:
[…]
