### 2025-06-01 d64f7de
fix: reducing number of words
