+++ paper/paper.md
@@ -56,9 +56,7 @@ The symbolic DA combines DA with symbolic calculation. Any coefficient of a Symb
 # Statement of need
 DA was used in particle beam dynamic analysis in the 1980s and was gradually extended to other fields. DA provides powerful analyzing tools, *i.e.*, map generation,  symplectic tracking [@caprimap], and normal form analysis [@monthnf], for a dynamic system. It is also used in verified integration [@rdaint], global optimization [@makino2005verified], and fast multipole method [@FMMCPO2010].   DA tools are available in several popular programs for particle accelerator design and simulations, *i.e.*,  COSY Infinity [@COSYCAP04], MAD-X [@grote2003mad; @MADX], and PTC [@forest2002introduction]. Stand-alone DA/TPSA libraries include DACE [@massari2018differential; @DACE] and cppTPSA/pyTPSA [@zhang2024cpptpsa]. All of them only perform numerical DA calculation. This library is the first and currently the only one that carries out symbolic DA calculations. 
 
-We developed SDA to improve the efficiency of repetitive DA processes. In some scenarios, we need to implement a complicated DA method repetitively on the same object with different initial conditions, *i.e.*, different positions and velocities of particles.  Some DA calculations are relatively slow. For example, calculating the inverse of a DA vector includes iterations and the number of iterations is determined by the order of the DA vector. This limits the use of DA in high-efficiency simulations. Using numerical DA, we lose track of the initial conditions in the calculation and have to repeat the same DA calculation for each case. Using symbolic DA, we only need to carry out the DA calculation once to obtain the final result as explicit expressions of the initial conditions. We then use those expressions on all the different initial conditions. By reducing the calculation on DA vectors to calculation on numbers, we can significantly enhance the efficiency. 
-
-SDA can also be used in higher-order derivative calculation. DA calculates the Taylor expansion of an arbitrary function. The partial derivatives of that function can be extracted from the coefficients of its Taylor expansion. Numerical DA calculates all the derivatives up to a predetermined order at a specific point. It is not possible to calculate a specific derivative without calculating the others. SDA provides explicit expressions for all the derivatives, allowing to calculate any specific derivative at any point. Calculating the derivatives using the expressions from SDA is usually faster. If only a few derivatives are needed, SDA can save even more time by avoiding redundant calculations on unnecessary derivatives.
+We developed SDA to improve the efficiency of repetitive DA processes. By performing the computation once with SDA, we obtain an explicit expression for how the final DA vector depends on the initial inputs. Evaluating this expression—rather than rerunning the full DA process for each new input—significantly reduces computation time  [@zhang2025Boosting]. SDA also provides an efficient method for computing higher‐order derivatives of any given function.
 
 # Features
 
