+++ paper/paper.bib
@@ -86,7 +86,7 @@ year        = "2006"}
 }
 
 @techreport{deniau2017mad,
-  title={The {MAD-X} Program, version 5.07.00, User’s Reference Manual},
+  title={The {MAD-X} Program, version 5.07.00, User\'s Reference Manual},
   author={Deniau, Laurent and Grote, Hans and Roy, Ghislain and Schmidt, Frank},
   institution={{CERN}},
   year={2017}
@@ -149,4 +149,13 @@ year        = "2006"}
   url          = {https://github.com/symengine/symengine},
   version      = {v0.12.0},
   note         = {Accessed: 2024-09-13}
+}
+
+@booklet{zhang2025Boosting,
+  author = {Zhang, He,
+  title = {Boosting the Efficiency of the Differential AlgebraBased Fast Multipole Method Operators Using Symbolic Calculation},
+  howpublished= {Presented at the meeting of SIAM CSE25, Fort Worth, Texas, U.S.},
+  year = {2025},
+  url = {https://www.siam.org/media/fyvh3qlf/cse25_abstracts.pdf},
+  note = {Accessed: 2025-06-01}
 }
