+++ paper/paper.bib
@@ -133,7 +133,7 @@ year        = "2006"}
 
 @article{zhang2024cpptpsa,
   title={{cppTPSA/pyTPSA}: a {C++/Python} package for truncated power series algebra},
-  author={Zhang, He},
+  author={Zhang, H.},
   journal={Journal of Open Source Software},
   volume={9},
   number={94},
@@ -152,7 +152,7 @@ year        = "2006"}
 }
 
 @booklet{zhang2025Boosting,
-  author = {Zhang, He,
+  author = {Zhang, H.},
   title = {Boosting the Efficiency of the Differential AlgebraBased Fast Multipole Method Operators Using Symbolic Calculation},
   howpublished= {Presented at the meeting of SIAM CSE25, Fort Worth, Texas, U.S.},
   year = {2025},
