+++ paper/paper.md
@@ -56,7 +56,7 @@ The symbolic DA combines DA with symbolic calculation. Any coefficient of a Symb
 # Statement of need
 DA was used in particle beam dynamic analysis in the 1980s and was gradually extended to other fields. DA provides powerful analyzing tools, *i.e.*, map generation,  symplectic tracking [@caprimap], and normal form analysis [@monthnf], for a dynamic system. It is also used in verified integration [@rdaint], global optimization [@makino2005verified], and fast multipole method [@FMMCPO2010].   DA tools are available in several popular programs for particle accelerator design and simulations, *i.e.*,  COSY Infinity [@COSYCAP04], MAD-X [@grote2003mad; @MADX], and PTC [@forest2002introduction]. Stand-alone DA/TPSA libraries include DACE [@massari2018differential; @DACE] and cppTPSA/pyTPSA [@zhang2024cpptpsa]. All of them only perform numerical DA calculation. This library is the first and currently the only one that carries out symbolic DA calculations. 
 
-We developed SDA to improve the efficiency of repetitive DA processes. By performing the computation once with SDA, we obtain an explicit expression for how the final DA vector depends on the initial inputs. Evaluating this expression—rather than rerunning the full DA process for each new input—significantly reduces computation time  [@zhang2025Boosting]. SDA also provides an efficient method for computing higher‐order derivatives of any given function.
+We developed SDA to improve the efficiency of repetitive DA processes. By performing the computation once with SDA, we obtain an explicit expression for how the final DA vector depends on the initial inputs. Evaluating this expression—rather than rerunning the full DA process for each new input—significantly reduces computation time [@zhang2025Boosting]. SDA also provides an efficient method for computing higher‐order derivatives of any given function.
 
 # Features
 
