### 2025-06-01 f73fabc
fix: shorten the paper
