+++ paper/paper.md
@@ -25,9 +25,7 @@ The truncated power series algebra (TPSA), or differential algebra (DA), is a we
 
 # Background
 
-We start with a brief introduction to DA from a practical computational perspective. Please refer to @AIEP108book and @chao2002lecture  for the complete theory.
-
-The fundamental concept in DA is the DA vector. To make this concept easier to understand, we can consider a DA vector as the Taylor expansion of a function at a specific point.  
+The fundamental concept in DA is the DA vector, which can be considered as the Taylor expansion of a function at a specific point from a practical  perspective.
 
 Considering a function $f(\mathbf{x})$ and its Taylor expansion $f_{\mathrm{T}}(\mathbf{x}_0)$  at the point $\mathbf{x}_0$ up to order $n$, we can define  the equivalence relation between the Taylor expansion and the DA vector as follows
 
@@ -41,20 +39,12 @@ c\cdot[a]_{n} & := & [c\cdot a]_{n},\label{eq:addmlt}\\
 
 According to the fixed point theorem  [@AIEP108book], the inverse of a DA vector that is not infinitely small can be calculated in a finite number of iterations. 
 
-The derivation operator $\partial_v$ with respect to the $v^{\mathrm{th}}$ variable can be defined as 
-
-$$ \partial_v[a]_n = \left[ \frac{\partial}{\partial x_v} a \right]_{n-1}, $$
-
-which can be carried out on a term by term basis on $[a]_n$. The operator $\partial_v$ satisfies the chain rule:
-
-$$ \partial_v([a]\cdot [b]) = [a]\cdot (\partial_v [b]) + (\partial_v [a])\cdot [b]. $$
-
-The inverse operator $\partial^{-1}_v$ can also be defined and applied easily on a term-by-term basis. Once the fundamental operators are defined, the DA vector can be used in calculations just as a number.
+The derivation operator $\partial_v$ with respect to the $v^{\mathrm{th}}$ variable and its inverse operator $\partial^{-1}v$ can be defined and carried out on a term by term basis on $[a]n$. Once the fundamental operators are defined, the DA vector can be used in calculations just as a number.
 
 The symbolic DA combines DA with symbolic calculation. Any coefficient of a Symbolic DA (SDA) vector is an explicit expression of the symbols in lieu of a number.  
 
 # Statement of need
-DA was used in particle beam dynamic analysis in the 1980s and was gradually extended to other fields. DA provides powerful analyzing tools, *i.e.*, map generation,  symplectic tracking [@caprimap], and normal form analysis [@monthnf], for a dynamic system. It is also used in verified integration [@rdaint], global optimization [@makino2005verified], and fast multipole method [@FMMCPO2010].   DA tools are available in several popular programs for particle accelerator design and simulations, *i.e.*,  COSY Infinity [@COSYCAP04], MAD-X [@grote2003mad; @MADX], and PTC [@forest2002introduction]. Stand-alone DA/TPSA libraries include DACE [@massari2018differential; @DACE] and cppTPSA/pyTPSA [@zhang2024cpptpsa]. All of them only perform numerical DA calculation. This library is the first and currently the only one that carries out symbolic DA calculations. 
+DA was used in particle beam dynamic analysis in the 1980s and was gradually extended to other fields. DA provides powerful analyzing tools [@caprimap; @monthnf], for a dynamic system. It is also used in various numerical algorithms [@rdaint; @makino2005verified; @FMMCPO2010].   DA tools are available in particle accelerator simulators [@COSYCAP04; @grote2003mad; @MADX; @forest2002introduction], or as stand-alone libraries  [@massari2018differential; @DACE; @zhang2024cpptpsa]. All of them only perform numerical DA calculation. This library is the first and currently the only one that carries out symbolic DA calculations. 
 
[…]
