### 2025-06-01 f219906
fix: shorten a few more words
