### 2025-06-02 #8330 @editorialbot
[REVIEW]: SDA: a symbolic differential algebra package in C++
**Submitting author:** @zhanghe9704 (<a href="http://orcid.org/0000-0001-7701-4118">He Zhang</a>)
**Repository:** https://github.com/zhanghe9704/tpsa_sym
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
