### 2025-06-03 a25bfbc
doc: add a reference to paper.md
