+++ paper/paper.bib
@@ -159,3 +159,13 @@ year        = "2006"}
   url = {https://www.siam.org/media/fyvh3qlf/cse25_abstracts.pdf},
   note = {Accessed: 2025-06-01}
 }
+
+@misc{zhang2025higher,
+      title={Higher-Order Automatic Differentiation Using Symbolic Differential Algebra: Bridging the Gap between Algorithmic and Symbolic Differentiation}, 
+      author={Zhang, H.},
+      year={2025},
+      eprint={2506.00796},
+      archivePrefix={arXiv},
+      primaryClass={physics.comp-ph},
+      url={https://arxiv.org/abs/2506.00796}, 
+}
