+++ paper/paper.bib
@@ -167,5 +167,5 @@ year        = "2006"}
       eprint={2506.00796},
       archivePrefix={arXiv},
       primaryClass={physics.comp-ph},
-      url={https://arxiv.org/abs/2506.00796}, 
+      url={https://arxiv.org/abs/2506.00796}
 }
