### 2025-06-23 #8330 @kpeeters
My main concern with this contribution is the lack of a clear "statement of need". The reference cited for this (Zhang, H. (2025)) is only an abstract in conference proceedings, with essentially the same announcement as in the JOSS paper itself. The very least that I was expecting is a clear explanation somewhere of how the method that is described here (namely doing the formal steps of differential algebra using symbolic rather than numeric expressions for (function value, derivative value)) compares, at the end of the day, to simply doing direct symbolic differentiation of the full expression.

I also think it lacks a proper documentation of the examples (they contain no comments, not even about the meaning of symbols used) and/or a tutorial-style documentation  (in my opinion the doxygen documentation is too low-level to be useful as a set of beginner instructions).
