### 2025-06-25 #8330 @zhanghe9704
@danielskatz @kpeeters

Thanks for reviewing my paper and code and giving me your feedback. I'll give my answer in the following.

- Statement of need: The SDA code will be useful in the following three cases.

1. Improving the efficiency of DA calculations. I developed a DA-based fast multipole method to calculate the Coulomb potential/field between charged particles, but it suffers from lower efficiency when compared with other FMM implementations. This problem triggered me to develop this SDA package. Using SDA, I can obtain the final expression after complicated DA calculations and generate high-efficiency code that avoids redundant calculation in numerical DA computation. I cannot rule out the possibility of carrying out all the DA calculations using a symbolic library, but I expect a significant effort would be needed to make it happen. However, using SDA, that can be done in just a few lines of code, exactly as one can do using a numerical DA package, because one can treat the expansions, i.e., DA variables, just like numbers. Alternatively, you can think of it this way: the SDA package takes care of the tedious work of "carrying out the DA calculation using a symbolic library" for you. The reference is an abstract for the SIAM CSE'25 conference. I presented a poster about preliminary results of the above work: we have achieved 20- to 50-fold speedup for DA-based FMM operators. But the conference does not have a proceeding. I am writing a paper on my latest results. It will probably take another two weeks for me to finalize the draft and go through the internal review process in my institute before I can put it on arXiv.

2. Dynamic study. If we are interested in the dynamic property of a particle accelerator, we can obtain the map of the accelerator using DA. Then we can do some analysis or tracking study. That is a routine task for accelerator physicists. If we want to know the effect of the strength of a specific magnet, we probably would adjust the value and repeat the analysis. But SDA will allow us to obtain the map as a function of the strength of that magnet, which would probably show in a more direct way how it contributes to the dynamics. Another benefit is a self-evolved map by SDA. For example, in a ramping process, the particles are accelerated and the setting of many elements will change accordingly. Numerical DA only gives a map for a fixed setting. SDA can give a map that changes with variables. I cannot give references on these because, as of now, they have only existed in my mind. These are just two possible implementations in accelerator physics. DA has been used in other fields, in which SDA may bring new insights.

3. Higher-order differentiation. This is a by-product. Derivative calculation has never been the main purpose of any DA package. But DA can be used as a tool for algorithmic/automatic differentiation. What makes SDA unique is that one can obtain explicit expressions of the derivatives following the algorithmic differentiation process instead of symbolic differentiation. I just finished a paper on this about two weeks ago: [https://arxiv.org/abs/2506.00796](https://arxiv.org/abs/2506.00796).

I could provide a long list of numerical DA implementations. Unfortunately, I cannot provide more references for SDA because this package is the first one to make symbolic DA calculation practically possible and it has not yet been exposed to a broad audience. Nonetheless, I believe it will prove useful to researchers who have been working with numerical DA packages.

I will revise the statement of need. However, I may not be able to include all the discussions above due to the word limit of 1,000.

- Documentation: I'll add instructions about how to use the code in the README.md and add comments in the examples.

I'll update the paper and the code and let you know when I have finished it. Thanks again for your effort.
