### 2025-06-25 #8330 @kpeeters
@zhanghe9704 Just as a general comment in response to the above: I would steer clear of mentioning too many physics applications. That will make your paper *less* accessible because you hide the real issues behind things that a large fraction of your potential audience is not interested in. Everyone reading this paper will know what differentiation is and why being able to do that by machine is useful. Most readers will be familiar with numerical automatic differentiation (AD). I think the point you want to emphasise is that your library is primarily useful for people who *already* have code in which some physics process is modeled using numerical AD, but for which the mathematical expressions are not easily written down in closed form and then differentiated symbolically (because it's all hidden in nested function calls). Your library then provides the tool to nevertheless get a *symbolic* expression for the derivatives (and which may then be faster to evaluate; if you had a simple example where this is true, that would substantially help to drive your point  home).

(Or in other words: If I had some complicated function in closed form and wanted derivatives, I would just stick it into Mathematica/SymPy/... and take the expression from there. Nobody would use SDA for that. You would only use SDA if that function is encoded in a complicated program with nested function calls, and manually constructing derivatives of that is tedious. This is what I was missing from your paper and documentation).

Hope this helps (just trying to help you get a larger audience; I think the result is nice and the code is definitely useful).
