### 2025-07-01 b181182
doc: add comments in examples
