### 2025-07-01 ad8c517
doc: add a short instruction on SDA usage
