+++ README.md
@@ -164,6 +164,112 @@ The simplest way to compile SDA in Windows is to use WSL2 and compile it as in L
 ### macOS ###
 Sorry! The author has never used macOS and hence cannot provide an instruction. However, if you can install SymEngine following its instruction successfully, compiling SDA lib is expected to be as straightforward as in Linux. 
 
+### How to use SDA ###
+We assume the users have some basic concepts of DA. The difference between DA and SDA lies in the fact that SDA allow you to use symbols in your DA calculation. 
+
+Similar to DA, before starting any SDA calculation, one needs to define the number of bases and the cut-off order of the SDA variable and the size of the memory pool that saves the SDA vectors. 
+in the following code, a memory pool that holds 400 SDA vectors of three bases up to the fourth order is created. 
+```C++
+    // Set DA environment parameters
+    int da_order = 4;   // Maximum order of DA expansion
+    int da_dim = 3;     // Number of variables (dimensions)
+    int da_pool = 400;  // Pool size to whole 400 DA vectors
+    da_init(da_order, da_dim, da_pool); // Initialize the DA environment
+```
+After calling `da_init()`, the bases are available as `da[i]` where i is 0 - 2. Now we can construct our first SDA vector using the bases. In the following code, we first defined three symbols, x, y, and z. Then we defined a first order SDA vector and the three symbols are used in the coefficients. 
+
+```C++
+    // Define symbolic variables using SymEngine
+    SymEngine::Expression x("x");
+    SymEngine::Expression y("y");
+    SymEngine::Expression z("z");
+
+    // Construct a symbolic DA vector using the symbolic variables
+    SymbDA::DAVector da1 = 1+(1+x)*da[0] + y*da[1] + (z-0.5)*da[2];
+```
+
+If we output the SDA vector `da1` to the screen, we will see as follows. A full 3D 4-order SDA vector will have 35 elements. `da1` has only 4 elements, including one zero-order term and three first-order terms. 
+
+```shell
+ Base      I        V [8]               [ 4 / 35 ]
+------------------------------------------------
+ 0 0 0     0    1.0
+ 1 0 0     1    1 + x
+ 0 1 0     2    y
+ 0 0 1     3    -0.5 + z
+```
+Once a SDA vector is defined, you can use it as a number. In the following codes, we define another SDA  vector `da2` and take the summation of both SDA vectors.  
+```C++
+SymbDA::DAVector da2 = 3.3+(0.5+x)*da[0] + y*y*da[1] + (x+z+1.1)*da[2];
+auto da3 = da1 + da2;
+```
+The value of `da2` is as follows. 
+```shell
+ Base      I        V [14]               [ 4 / 35 ]
+------------------------------------------------
+ 0 0 0     0    3.3
+ 1 0 0     1    0.5 + x
+ 0 1 0     2    y**2
+ 0 0 1     3    1.1 + x + z
+```
[…]
