### 2025-07-01 7f1fd8a
doc: fix build-examples/tests to build_examples/tests
