### 2025-07-01 77d01a0
doc: revise a little here and there
