### 2025-07-01 #8330 @zhanghe9704
@danielskatz @kpeeters Added a short instruction on SDA calculation in README.md and added comments in the examples. I'll update the paper tomorrow.
