### 2025-07-02 2a5df10
doc: revise statement of need
