### 2025-07-02 0ddbdea
doc: shorten the paper
