### 2025-07-02 #8330 @zhanghe9704
@danielskatz @kpeeters

I updated the repository:
1. Add a short instruction on the usage of SDA in README.md.
2. Add comments on the examples.
3. Revise the statement of need.

Thanks to @kpeeters  for your suggestions. I fully agree with the following two points:
1. Do not include specific physics examples.
2. The target users are people who have experience with DA.

I think I also need to clear a concept: DA, also referred to as TPSA (truncated power series algorithm) is not the AD that is widely used in machine learning. AD calculates the first-order derivative. For higher-order derivatives, one needs a nested AD approach. When the order goes up, e.g., above 5, the nested structure could be very slow. DA/TPSA is based on power series and fixed-point theory. It is originally designed for dynamic analysis, in which high-order derivatives are often needed. DA/TPSA calculates high-order Taylor expansion in a very efficient way without nesting.

Should we only use SDA when we do not have the explicit expression? I think it is case by case. For the example of DA-based FMM, we do have explicit expressions of the FMM operators. But they are very complex, including multiplication and inversion of high-order (up to 10) 3D full rank Cartesian tensors and the composition of them, considering that a DA vector is also a Cartesian tensor. I believe using any symbolic lib to derive the final expression will be very complicated, although possible. But since we can express the operators in the DA framework, SDA can easily get the job done. Actually, I'm curious if this trick can be used on other tensor-based or power series-based computations.

Higher-order differentiation is not the main purpose of DA/SDA package. I mention it because I feel it is CONCEPTUALLY interesting that SDA allows us to obtain symbolic expression via an algorithmic differentiation process rather than a normal symbolic differentiation process. People usually think that algorithmic differentiation and symbolic differentiation are two separate ways, but here they are connected. I don't know how SDA compares with other symbolic libs on performance. I would like to see someone compare them, but it is not my current priority. BTW, the symbolic calculation in SDA is done by SymEngine, which is usually considered a good lib for symbolic calculation.

I didn't include the above discussion in the paper because I think they are out of scope. I wrote them as general discussions on DA/SDA and wish to clear up misunderstanding.

Thanks again!
