### 2025-07-07 1762533
fix: static variable da cause memory error in termination
