+++ paper/paper.md
@@ -42,9 +42,9 @@ According to the fixed point theorem  [@AIEP108book], the inverse of a DA vector
 The symbolic DA combines DA with symbolic calculation. Any coefficient of a Symbolic DA (SDA) vector is an explicit expression of the symbols in lieu of a number.  
 
 # Statement of need
-DA has been used in particle beam dynamic analysis since1980s and gradually extended to other fields. DA provides powerful analyzing tools [@caprimap; @monthnf], for a dynamic system. It is also used in various numerical algorithms， *e.g.* verified integration  [@rdaint], global optimization [@makino2005verified], and fast multipole method [@FMMCPO2010].   DA tools are available in particle accelerator simulators [@COSYCAP04; @grote2003mad; @MADX; @forest2002introduction], or as stand-alone libraries  [@massari2018differential; @DACE; @zhang2024cpptpsa]. All of them only perform numerical DA calculation. This is the first and only library for symbolic DA calculations. 
+DA has been used in particle beam dynamic analysis since the 1980s and gradually extended to other fields. DA provides powerful analyzing tools [@caprimap; @monthnf], for analyzing dynamic systems. It is also used in various numerical algorithms, *e.g.*, verified integration  [@rdaint], global optimization [@makino2005verified], and fast multipole method [@FMMCPO2010].   DA tools are available in particle accelerator simulators [@COSYCAP04; @grote2003mad; @MADX; @forest2002introduction], or as stand-alone libraries  [@massari2018differential; @DACE; @zhang2024cpptpsa]. All of them only perform numerical DA calculations. This is the first and only library for symbolic DA calculations. 
 
-The SDA package can be used in several important ways. First, in high-performance computing scenarios involving differential algebra (DA), SDA can produce explicit expressions for the results of complex DA calculations. These expressions enable the generation of highly efficient code that replaces runtime DA computations, resulting in significant performance improvements [@zhang2025Boosting]. This was the original motivation for developing the package. Second, as a tool for dynamic analysis, SDA supports parametric and time-dependent studies by allowing DA maps to vary with symbolic parameters. This makes it possible to analyze how system behavior changes with evolving settings—something difficult to achieve with numerical DA alone. Finally, SDA provides a new way to obtain symbolic expressions of higher-order derivatives by following the algorithmic differentiation process, effectively bridging the gap between symbolic and numerical approaches  [@zhang2025higher].
+The SDA package can be used in several important ways. First, in high-performance computing scenarios involving differential algebra (DA), it can produce explicit expressions for the results of complex DA calculations. These expressions enable the generation of highly efficient code that eliminates the need for runtime DA computations, resulting in significant performance improvements [@zhang2025Boosting]. This was the original motivation for developing the package. Second, as a tool for dynamic analysis, it supports parametric and time-dependent studies by allowing DA maps to vary with symbolic parameters. This makes it possible to analyze how system behavior changes with evolving settings—something difficult to achieve with numerical DA alone. Finally, it provides a new way to obtain symbolic expressions of higher-order derivatives by following the algorithmic differentiation process, thus bridging the gap between symbolic and numerical approaches  [@zhang2025higher].
 
 
 
