### 2025-07-07 #8330 @zhanghe9704
@danielskatz @mbkumar

After installing the Intel Open API toolkit, I was able to reproduce the  Aborted (core dumped) error and fix it. I have also edited the statement of need. Replied on the Issues with a little more detail. Thanks a lot!
