### 2025-07-10 #8330 @danielskatz
@mbkumar - Are you able to continue now?
