### 2025-07-10 #8330 @mbkumar
Yes. Will get to it sometime today

On Thu, Jul 10, 2025, 9:10 AM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#8330)
> <https://github.com/openjournals/joss-reviews/issues/8330#issuecomment-3057396523>
>
> @mbkumar <https://github.com/mbkumar> - Are you able to continue now?
>
