### 2025-08-05 #8330 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
